[refactor] Scanner를 받지 않는 기본 생성자로 변경
