+++ src/main/java/step3/controller/LottoController.java
@@ -1,8 +1,8 @@
 package step3.controller;
 
 import step3.domain.lotto.LottoTicketMachine;
-import step3.domain.lotto.LottoTickets;
 import step3.domain.lotto.WinningNumbers;
+import step3.domain.lotto.firstcollection.LottoTickets;
 import step3.strategy.LottoNumberMakeStrategy;
 import step3.view.InputView;
 import step3.view.LottoInputView;
@@ -16,7 +16,7 @@ public class LottoController {
     private final ResultView resultView;
 
     public LottoController() {
-        this(new LottoInputView(), new LottoResultView());
+        this(new LottoInputView(new Scanner(System.in)), new LottoResultView());
     }
 
     public LottoController(InputView inputView, ResultView resultView) {
@@ -25,13 +25,12 @@ public class LottoController {
     }
 
     public void ticketing() {
-        Scanner scanner = new Scanner(System.in);
-        int useAmount = inputView.getUseAmount(scanner);
+        int useAmount = inputView.getUseAmount();
         LottoTickets tickets = LottoTicketMachine.ticketing(useAmount, new LottoNumberMakeStrategy());
 
         resultView.drawTicket(tickets);
 
-        WinningNumbers winningNumbers = inputView.getWinningNumber(scanner);
+        WinningNumbers winningNumbers = inputView.getWinningNumber();
         resultView.drawWinningStatistics(tickets, winningNumbers);
         resultView.drawRevenueRate(tickets, winningNumbers);
     }
