+++ src/main/java/step3/view/InputView.java
@@ -2,9 +2,7 @@ package step3.view;
 
 import step3.domain.lotto.WinningNumbers;
 
-import java.util.Scanner;
-
 public interface InputView {
-    int getUseAmount(Scanner scanner);
-    WinningNumbers getWinningNumber(Scanner scanner);
+    int getUseAmount();
+    WinningNumbers getWinningNumber();
 }
